requirements.txt
click>=8.1.3
flask==3.0.2
fastapi==0.65.1
uvicorn>=0.14.0
gspread==6.0.2
oauth2client==4.1.3
google-cloud-secret-manager